% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saeHB.ME.beta.R
\docType{package}
\name{saeHB.ME.beta}
\alias{saeHB.ME.beta}
\title{saeHB.ME.beta: Small Area Estimation with Measurement Error using Hierarchical Bayesian Method under Beta Distribution}
\description{
Implementation of small area estimation using Hierarchical Bayesian (HB) Method when auxiliary variable measured with error under Beta Distribution.
The 'rjags' package is employed to obtain parameter estimates.
}
\section{Authors}{

Azka Ubaidillah, Ratih Rodliyah
}

\section{Email}{

Ratih Rodliyah \email{ratihrodliyah2@gmail.com}
}

\section{Functions}{

\describe{
\item{\code{\link{meHBbeta}}}{Produces HB estimators, standard error, random effect variance, coefficient and plot under beta distribution.}
}
}

\references{
Rao, J.N.K & Molina. (2015). Small Area Estimation 2nd Edition. New York: John Wiley and Sons, Inc \url{doi:10.1002/9781118735855}.

Ybarra, L.M. and Lohr, S. L. (2008). Small area estimation when auxiliary information is measured with error. Biometrika 95, 919-931 \url{doi:10.1093/biomet/asn048}.

Ntzoufras, I. (2009), Bayesian Modeling Using WinBUGS. 1st Edn., Wiley, New Jersey, ISBN-10: 1118210352.
}
